requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.1
websockets==12.0
celery==5.3.4
redis==5.0.1
python-magic==0.4.27
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
